Keep the original websites sort order
This way the user can decide which website has higher priority if two or
more websites would match.

e.g. two websites with the same host, one matched byHost and the other
byUserParam, would require that the website matched byUserParam is set
earlier than the other to stand a chance.